Prepare score table for decoys

Takes an input object and returns a score table for the decoys.

decoyScoreTable(object, decoy, score, log10 = TRUE)

Arguments

object

A data.frame, mzID or mzRident object.

decoy

character, name of the variable that indicates if the peptide matches to a target or to a decoy. When this value is missing, a Shiny gadget is launched to select it interactively.

score

numeric, indicating the score of the peptide match, obtained by the search engine. When this value is missing, a Shiny gadget is launched to select it interactively.

log10

logical to indicate if the score should be -log10-transformed. Default: TRUE. When this value is missing, a Shiny gadget is launched to select it interactively.

Value

A data.frame with a logical "decoy" column and numeric "scores".

Author

Elke Debrie, Lieven Clement

Examples

library(mzID)

## Use one of the example files in the mzID package
exampleFile <- system.file("extdata", "55merge_tandem.mzid", package = "mzID")
mzIDexample <- mzID(exampleFile)
#> reading 55merge_tandem.mzid... DONE!

decoyScoreTable(mzIDexample, decoy = "isdecoy", score = "x\\!tandem:expect")
#>       decoy       score
#> 1     FALSE  3.76955108
#> 2     FALSE  2.53760200
#> 3     FALSE  2.52287875
#> 4     FALSE  2.29242982
#> 5     FALSE  1.79588002
#> 6      TRUE  1.17392520
#> 7      TRUE  1.00000000
#> 8      TRUE  0.88605665
#> 9      TRUE  0.79588002
#> 10     TRUE  0.76955108
#> 11     TRUE  0.74472749
#> 12     TRUE  0.65757732
#> 13     TRUE  0.60205999
#> 14    FALSE  0.58502665
#> 15     TRUE  0.56863624
#> 16     TRUE  0.48148606
#> 17     TRUE  0.45593196
#> 18     TRUE  0.45593196
#> 19    FALSE  0.45593196
#> 20    FALSE  0.43179828
#> 21     TRUE  0.38721614
#> 22     TRUE  0.37675071
#> 23     TRUE  0.36653154
#> 24     TRUE  0.33724217
#> 25     TRUE  0.29242982
#> 26     TRUE  0.29242982
#> 27     TRUE  0.26760624
#> 28     TRUE  0.25963731
#> 29    FALSE  0.25963731
#> 30    FALSE  0.24412514
#> 31     TRUE  0.19382003
#> 32     TRUE  0.19382003
#> 33     TRUE  0.18708664
#> 34     TRUE  0.14874165
#> 35     TRUE  0.13667714
#> 36    FALSE  0.10790540
#> 37     TRUE  0.10237291
#> 38     TRUE  0.08092191
#> 39     TRUE  0.04575749
#> 40     TRUE  0.04575749
#> 41     TRUE  0.03151705
#> 42     TRUE  0.01322827
#> 43    FALSE -0.04139269
#> 44     TRUE -0.04139269
#> 45     TRUE -0.04139269
#> 46     TRUE -0.04139269
#> 47     TRUE -0.04139269
#> 48     TRUE -0.04139269
#> 49     TRUE -0.04139269
#> 50    FALSE -0.04139269
#> 51    FALSE -0.07918125
#> 52    FALSE -0.07918125
#> 53     TRUE -0.07918125
#> 54     TRUE -0.07918125
#> 55     TRUE -0.07918125
#> 56    FALSE -0.07918125
#> 57     TRUE -0.07918125
#> 58     TRUE -0.07918125
#> 59    FALSE -0.07918125
#> 60    FALSE -0.11394335
#> 61     TRUE -0.11394335
#> 62     TRUE -0.11394335
#> 63     TRUE -0.11394335
#> 64    FALSE -0.11394335
#> 65     TRUE -0.11394335
#> 66     TRUE -0.14612804
#> 67     TRUE -0.14612804
#> 68     TRUE -0.14612804
#> 69     TRUE -0.14612804
#> 70     TRUE -0.17609126
#> 71     TRUE -0.17609126
#> 72     TRUE -0.17609126
#> 73     TRUE -0.17609126
#> 74     TRUE -0.20411998
#> 75     TRUE -0.20411998
#> 76     TRUE -0.23044892
#> 77    FALSE -0.25527251
#> 78     TRUE -0.25527251
#> 79    FALSE -0.27875360
#> 80     TRUE -0.27875360
#> 81    FALSE -0.27875360
#> 82     TRUE -0.30103000
#> 83    FALSE -0.30103000
#> 84     TRUE -0.32221929
#> 85     TRUE -0.32221929
#> 86    FALSE -0.34242268
#> 87     TRUE -0.34242268
#> 88     TRUE -0.34242268
#> 89    FALSE -0.34242268
#> 90     TRUE -0.34242268
#> 91     TRUE -0.34242268
#> 92     TRUE -0.36172784
#> 93     TRUE -0.36172784
#> 94    FALSE -0.36172784
#> 95    FALSE -0.36172784
#> 96    FALSE -0.38021124
#> 97     TRUE -0.38021124
#> 98     TRUE -0.38021124
#> 99     TRUE -0.38021124
#> 100    TRUE -0.39794001
#> 101    TRUE -0.39794001
#> 102    TRUE -0.41497335
#> 103   FALSE -0.41497335
#> 104    TRUE -0.41497335
#> 105    TRUE -0.41497335
#> 106   FALSE -0.43136376
#> 107    TRUE -0.43136376
#> 108    TRUE -0.44715803
#> 109   FALSE -0.44715803
#> 110    TRUE -0.46239800
#> 111    TRUE -0.46239800
#> 112    TRUE -0.46239800
#> 113    TRUE -0.47712125
#> 114    TRUE -0.47712125
#> 115    TRUE -0.47712125
#> 116    TRUE -0.49136169
#> 117    TRUE -0.49136169
#> 118    TRUE -0.50514998
#> 119    TRUE -0.53147892
#> 120   FALSE -0.53147892
#> 121    TRUE -0.53147892
#> 122   FALSE -0.54406804
#> 123    TRUE -0.54406804
#> 124   FALSE -0.55630250
#> 125   FALSE -0.55630250
#> 126    TRUE -0.56820172
#> 127   FALSE -0.56820172
#> 127.1  TRUE -0.56820172
#> 128    TRUE -0.57978360
#> 129    TRUE -0.59106461
#> 130   FALSE -0.60205999
#> 130.1 FALSE -0.60205999
#> 131    TRUE -0.61278386
#> 132    TRUE -0.61278386
#> 133    TRUE -0.62324929
#> 134   FALSE -0.62324929
#> 135    TRUE -0.63346846
#> 136    TRUE -0.65321251
#> 137    TRUE -0.65321251
#> 138    TRUE -0.66275783
#> 139    TRUE -0.67209786
#> 140   FALSE -0.67209786
#> 141   FALSE -0.69019608
#> 142    TRUE -0.69019608
#> 143    TRUE -0.69897000
#> 144    TRUE -0.69897000
#> 145    TRUE -0.70757018
#> 146   FALSE -0.72427587
#> 147    TRUE -0.72427587
#> 148    TRUE -0.73239376
#> 149    TRUE -0.74818803
#> 150    TRUE -0.74818803
#> 151   FALSE -0.77085201
#> 152    TRUE -0.77815125
#> 153    TRUE -0.77815125
#> 154    TRUE -0.78532984
#> 155    TRUE -0.79934055
#> 156   FALSE -0.80617997
#> 157    TRUE -0.80617997
#> 158   FALSE -0.81291336
#> 159    TRUE -0.81291336
#> 160    TRUE -0.81954394
#> 161    TRUE -0.83884909
#> 162    TRUE -0.83884909
#> 163   FALSE -0.84509804
#> 164    TRUE -0.84509804
#> 165    TRUE -0.85733250
#> 166   FALSE -0.86923172
#> 167    TRUE -0.88081359
#> 168    TRUE -0.89762709
#> 169    TRUE -0.97312785